BackEnd/requirements.txt
asgiref==3.6.0
config==0.5.1
Django==4.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.1
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
